feat: split admin dashboard into sections
